+++ Knihovna/ChangePasswordForm.Designer.cs
@@ -2,20 +2,13 @@
 {
     partial class ChangePasswordForm
     {
-        /// <summary>
-        /// Required designer variable.
-        /// </summary>
         private System.ComponentModel.IContainer components = null;
-        private System.Windows.Forms.Label lblNewPassword;
-        private System.Windows.Forms.Label lblConfirmPassword;
         private System.Windows.Forms.TextBox txtNewPassword;
         private System.Windows.Forms.TextBox txtConfirmPassword;
         private System.Windows.Forms.Button btnChangePassword;
+        private System.Windows.Forms.Label lblNewPassword;
+        private System.Windows.Forms.Label lblConfirmPassword;
 
-        /// <summary>
-        /// Clean up any resources being used.
-        /// </summary>
-        /// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
         protected override void Dispose(bool disposing)
         {
             if (disposing && (components != null))
@@ -25,80 +18,105 @@
             base.Dispose(disposing);
         }
 
-        #region Windows Form Designer generated code
-
-        /// <summary>
-        /// Required method for Designer support - do not modify
-        /// the contents of this method with the code editor.
-        /// </summary>
         private void InitializeComponent()
         {
-            this.lblNewPassword = new System.Windows.Forms.Label();
-            this.lblConfirmPassword = new System.Windows.Forms.Label();
             this.txtNewPassword = new System.Windows.Forms.TextBox();
             this.txtConfirmPassword = new System.Windows.Forms.TextBox();
             this.btnChangePassword = new System.Windows.Forms.Button();
+            this.lblNewPassword = new System.Windows.Forms.Label();
+            this.lblConfirmPassword = new System.Windows.Forms.Label();
+            this.labelUser = new System.Windows.Forms.Label();
             this.SuspendLayout();
+            // 
+            // txtNewPassword
+            // 
+            this.txtNewPassword.Font = new System.Drawing.Font("Segoe UI", 10F);
+            this.txtNewPassword.Location = new System.Drawing.Point(13, 82);
+            this.txtNewPassword.Margin = new System.Windows.Forms.Padding(4, 6, 4, 6);
+            this.txtNewPassword.Name = "txtNewPassword";
+            this.txtNewPassword.PasswordChar = '*';
+            this.txtNewPassword.Size = new System.Drawing.Size(170, 30);
+            this.txtNewPassword.TabIndex = 0;
+            // 
+            // txtConfirmPassword
+            // 
+            this.txtConfirmPassword.Font = new System.Drawing.Font("Segoe UI", 10F);
+            this.txtConfirmPassword.Location = new System.Drawing.Point(13, 153);
+            this.txtConfirmPassword.Margin = new System.Windows.Forms.Padding(4, 6, 4, 6);
+            this.txtConfirmPassword.Name = "txtConfirmPassword";
+            this.txtConfirmPassword.PasswordChar = '*';
+            this.txtConfirmPassword.Size = new System.Drawing.Size(170, 30);
+            this.txtConfirmPassword.TabIndex = 1;
+            // 
+            // btnChangePassword
+            // 
+            this.btnChangePassword.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(72)))), ((int)(((byte)(126)))), ((int)(((byte)(176)))));
+            this.btnChangePassword.FlatStyle = System.Windows.Forms.FlatStyle.Flat;
+            this.btnChangePassword.Font = new System.Drawing.Font("Segoe UI", 10F, System.Drawing.FontStyle.Bold);
+            this.btnChangePassword.ForeColor = System.Drawing.Color.White;
+            this.btnChangePassword.Location = new System.Drawing.Point(13, 197);
+            this.btnChangePassword.Margin = new System.Windows.Forms.Padding(4, 6, 4, 6);
+            this.btnChangePassword.Name = "btnChangePassword";
+            this.btnChangePassword.Size = new System.Drawing.Size(170, 35);
+            this.btnChangePassword.TabIndex = 2;
+            this.btnChangePassword.Text = "Změnit heslo";
+            this.btnChangePassword.UseVisualStyleBackColor = false;
+            this.btnChangePassword.Click += new System.EventHandler(this.btnChangePassword_Click);
             // 
             // lblNewPassword
             // 
             this.lblNewPassword.AutoSize = true;
-            this.lblNewPassword.Location = new System.Drawing.Point(12, 15);
+            this.lblNewPassword.Font = new System.Drawing.Font("Segoe UI", 10F);
+            this.lblNewPassword.Location = new System.Drawing.Point(13, 53);
+            this.lblNewPassword.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
             this.lblNewPassword.Name = "lblNewPassword";
-            this.lblNewPassword.Size = new System.Drawing.Size(87, 17);
-            this.lblNewPassword.TabIndex = 0;
+            this.lblNewPassword.Size = new System.Drawing.Size(99, 23);
+            this.lblNewPassword.TabIndex = 3;
             this.lblNewPassword.Text = "Nové heslo:";
             // 
             // lblConfirmPassword
             // 
             this.lblConfirmPassword.AutoSize = true;
-            this.lblConfirmPassword.Location = new System.Drawing.Point(12, 47);
+            this.lblConfirmPassword.Font = new System.Drawing.Font("Segoe UI", 10F);
+            this.lblConfirmPassword.Location = new System.Drawing.Point(13, 125);
+            this.lblConfirmPassword.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
             this.lblConfirmPassword.Name = "lblConfirmPassword";
-            this.lblConfirmPassword.Size = new System.Drawing.Size(118, 17);
-            this.lblConfirmPassword.TabIndex = 1;
+            this.lblConfirmPassword.Size = new System.Drawing.Size(128, 23);
+            this.lblConfirmPassword.TabIndex = 4;
             this.lblConfirmPassword.Text = "Potvrzení hesla:";
             // 
-            // txtNewPassword
+            // labelUser
             // 
-            this.txtNewPassword.Location = new System.Drawing.Point(136, 12);
-            this.txtNewPassword.Name = "txtNewPassword";
-            this.txtNewPassword.PasswordChar = '*';
-            this.txtNewPassword.Size = new System.Drawing.Size(200, 22);
-            this.txtNewPassword.TabIndex = 2;
-            // 
-            // txtConfirmPassword
-            // 
-            this.txtConfirmPassword.Location = new System.Drawing.Point(136, 44);
-            this.txtConfirmPassword.Name = "txtConfirmPassword";
-            this.txtConfirmPassword.PasswordChar = '*';
-            this.txtConfirmPassword.Size = new System.Drawing.Size(200, 22);
-            this.txtConfirmPassword.TabIndex = 3;
-            // 
-            // btnChangePassword
-            // 
-            this.btnChangePassword.Location = new System.Drawing.Point(136, 72);
-            this.btnChangePassword.Name = "btnChangePassword";
-            this.btnChangePassword.Size = new System.Drawing.Size(200, 30);
-            this.btnChangePassword.TabIndex = 4;
-            this.btnChangePassword.Text = "Změnit heslo";
-            this.btnChangePassword.UseVisualStyleBackColor = true;
-            this.btnChangePassword.Click += new System.EventHandler(this.btnChangePassword_Click);
+            this.labelUser.AutoSize = true;
+            this.labelUser.Font = new System.Drawing.Font("Segoe UI", 10F);
+            this.labelUser.Location = new System.Drawing.Point(9, 18);
+            this.labelUser.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
+            this.labelUser.Name = "labelUser";
+            this.labelUser.Size = new System.Drawing.Size(143, 23);
+            this.labelUser.TabIndex = 5;
+            this.labelUser.Text = "Změna hesla pro ";
             // 
             // ChangePasswordForm
             // 
-            this.ClientSize = new System.Drawing.Size(348, 114);
+            this.AutoScaleDimensions = new System.Drawing.SizeF(9F, 23F);
+            this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            this.BackColor = System.Drawing.Color.White;
+            this.ClientSize = new System.Drawing.Size(202, 252);
+            this.Controls.Add(this.labelUser);
+            this.Controls.Add(this.lblConfirmPassword);
+            this.Controls.Add(this.lblNewPassword);
             this.Controls.Add(this.btnChangePassword);
             this.Controls.Add(this.txtConfirmPassword);
             this.Controls.Add(this.txtNewPassword);
-            this.Controls.Add(this.lblConfirmPassword);
-            this.Controls.Add(this.lblNewPassword);
+            this.Font = new System.Drawing.Font("Segoe UI", 10F);
+            this.Margin = new System.Windows.Forms.Padding(4, 6, 4, 6);
             this.Name = "ChangePasswordForm";
             this.Text = "Změna hesla";
             this.ResumeLayout(false);
             this.PerformLayout();
+
         }
 
-        #endregion
-
+        private System.Windows.Forms.Label labelUser;
     }
 }