+++ Knihovna/LoginForm.Designer.cs
@@ -8,6 +8,8 @@
         private System.Windows.Forms.Button btnLogin;
         private System.Windows.Forms.Label lblUsername;
         private System.Windows.Forms.Label lblPassword;
+        private System.Windows.Forms.Panel panelHeader;
+        private System.Windows.Forms.Label labelNazev;
 
         protected override void Dispose(bool disposing)
         {
@@ -20,71 +22,131 @@
 
         private void InitializeComponent()
         {
+            this.panelHeader = new System.Windows.Forms.Panel();
+            this.labelNazev = new System.Windows.Forms.Label();
             this.txtUsername = new System.Windows.Forms.TextBox();
             this.txtPassword = new System.Windows.Forms.TextBox();
             this.btnLogin = new System.Windows.Forms.Button();
             this.lblUsername = new System.Windows.Forms.Label();
             this.lblPassword = new System.Windows.Forms.Label();
+            this.chybaPrihlaseniLabel = new System.Windows.Forms.Label();
+            this.panelHeader.SuspendLayout();
             this.SuspendLayout();
-
+            // 
+            // panelHeader
+            // 
+            this.panelHeader.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(72)))), ((int)(((byte)(126)))), ((int)(((byte)(176)))));
+            this.panelHeader.Controls.Add(this.labelNazev);
+            this.panelHeader.Dock = System.Windows.Forms.DockStyle.Top;
+            this.panelHeader.Location = new System.Drawing.Point(0, 0);
+            this.panelHeader.Name = "panelHeader";
+            this.panelHeader.Size = new System.Drawing.Size(450, 50);
+            this.panelHeader.TabIndex = 0;
+            // 
+            // labelNazev
+            // 
+            this.labelNazev.Font = new System.Drawing.Font("Segoe UI", 16F, System.Drawing.FontStyle.Bold);
+            this.labelNazev.ForeColor = System.Drawing.Color.Snow;
+            this.labelNazev.Location = new System.Drawing.Point(171, 6);
+            this.labelNazev.Name = "labelNazev";
+            this.labelNazev.Size = new System.Drawing.Size(141, 44);
+            this.labelNazev.TabIndex = 0;
+            this.labelNazev.Text = "Knihovna";
+            // 
             // txtUsername
-            this.txtUsername.Location = new System.Drawing.Point(16, 31);
-            this.txtUsername.Margin = new System.Windows.Forms.Padding(4);
+            // 
+            this.txtUsername.Font = new System.Drawing.Font("Segoe UI", 10F);
+            this.txtUsername.Location = new System.Drawing.Point(22, 108);
+            this.txtUsername.Margin = new System.Windows.Forms.Padding(4, 6, 4, 6);
             this.txtUsername.Name = "txtUsername";
-            this.txtUsername.Size = new System.Drawing.Size(345, 22);
-            this.txtUsername.TabIndex = 0;
-
+            this.txtUsername.Size = new System.Drawing.Size(404, 30);
+            this.txtUsername.TabIndex = 1;
+            // 
             // txtPassword
-            this.txtPassword.Location = new System.Drawing.Point(16, 79);
-            this.txtPassword.Margin = new System.Windows.Forms.Padding(4);
+            // 
+            this.txtPassword.Font = new System.Drawing.Font("Segoe UI", 10F);
+            this.txtPassword.Location = new System.Drawing.Point(22, 179);
+            this.txtPassword.Margin = new System.Windows.Forms.Padding(4, 6, 4, 6);
             this.txtPassword.Name = "txtPassword";
-            this.txtPassword.Size = new System.Drawing.Size(345, 22);
-            this.txtPassword.TabIndex = 1;
+            this.txtPassword.Size = new System.Drawing.Size(404, 30);
+            this.txtPassword.TabIndex = 2;
             this.txtPassword.UseSystemPasswordChar = true;
-
+            this.txtPassword.KeyPress += new System.Windows.Forms.KeyPressEventHandler(this.txtPassword_KeyPress);
+            // 
             // btnLogin
-            this.btnLogin.Location = new System.Drawing.Point(16, 111);
-            this.btnLogin.Margin = new System.Windows.Forms.Padding(4);
+            // 
+            this.btnLogin.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(72)))), ((int)(((byte)(126)))), ((int)(((byte)(176)))));
+            this.btnLogin.FlatStyle = System.Windows.Forms.FlatStyle.Flat;
+            this.btnLogin.Font = new System.Drawing.Font("Segoe UI", 10F, System.Drawing.FontStyle.Bold);
+            this.btnLogin.ForeColor = System.Drawing.Color.White;
+            this.btnLogin.Location = new System.Drawing.Point(22, 240);
+            this.btnLogin.Margin = new System.Windows.Forms.Padding(4, 6, 4, 6);
             this.btnLogin.Name = "btnLogin";
-            this.btnLogin.Size = new System.Drawing.Size(100, 28);
-            this.btnLogin.TabIndex = 2;
+            this.btnLogin.Size = new System.Drawing.Size(112, 43);
+            this.btnLogin.TabIndex = 3;
             this.btnLogin.Text = "Login";
-            this.btnLogin.UseVisualStyleBackColor = true;
+            this.btnLogin.UseVisualStyleBackColor = false;
             this.btnLogin.Click += new System.EventHandler(this.btnLogin_Click);
-
+            // 
             // lblUsername
+            // 
             this.lblUsername.AutoSize = true;
-            this.lblUsername.Location = new System.Drawing.Point(16, 11);
+            this.lblUsername.Font = new System.Drawing.Font("Segoe UI", 10F);
+            this.lblUsername.Location = new System.Drawing.Point(22, 79);
             this.lblUsername.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
             this.lblUsername.Name = "lblUsername";
-            this.lblUsername.Size = new System.Drawing.Size(70, 16);
-            this.lblUsername.TabIndex = 3;
+            this.lblUsername.Size = new System.Drawing.Size(87, 23);
+            this.lblUsername.TabIndex = 4;
             this.lblUsername.Text = "Username";
-
+            // 
             // lblPassword
+            // 
             this.lblPassword.AutoSize = true;
-            this.lblPassword.Location = new System.Drawing.Point(16, 59);
+            this.lblPassword.Font = new System.Drawing.Font("Segoe UI", 10F);
+            this.lblPassword.Location = new System.Drawing.Point(22, 151);
             this.lblPassword.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
             this.lblPassword.Name = "lblPassword";
-            this.lblPassword.Size = new System.Drawing.Size(43, 16);
-            this.lblPassword.TabIndex = 4;
-            this.lblPassword.Text = "Heslo";
-
+            this.lblPassword.Size = new System.Drawing.Size(80, 23);
+            this.lblPassword.TabIndex = 5;
+            this.lblPassword.Text = "Password";
+            // 
+            // chybaPrihlaseniLabel
+            // 
+            this.chybaPrihlaseniLabel.AutoSize = true;
+            this.chybaPrihlaseniLabel.Font = new System.Drawing.Font("Segoe UI", 10.2F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(238)));
+            this.chybaPrihlaseniLabel.ForeColor = System.Drawing.Color.Red;
+            this.chybaPrihlaseniLabel.Location = new System.Drawing.Point(22, 211);
+            this.chybaPrihlaseniLabel.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
+            this.chybaPrihlaseniLabel.Name = "chybaPrihlaseniLabel";
+            this.chybaPrihlaseniLabel.Size = new System.Drawing.Size(166, 23);
+            this.chybaPrihlaseniLabel.TabIndex = 6;
+            this.chybaPrihlaseniLabel.Text = "Neplatné přihlášení";
+            this.chybaPrihlaseniLabel.Visible = false;
+            // 
             // LoginForm
-            this.AutoScaleDimensions = new System.Drawing.SizeF(8F, 16F);
+            // 
+            this.AutoScaleDimensions = new System.Drawing.SizeF(9F, 23F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.ClientSize = new System.Drawing.Size(379, 149);
+            this.BackColor = System.Drawing.Color.White;
+            this.ClientSize = new System.Drawing.Size(450, 300);
+            this.Controls.Add(this.chybaPrihlaseniLabel);
             this.Controls.Add(this.lblPassword);
             this.Controls.Add(this.lblUsername);
             this.Controls.Add(this.btnLogin);
             this.Controls.Add(this.txtPassword);
             this.Controls.Add(this.txtUsername);
-            this.Margin = new System.Windows.Forms.Padding(4);
+            this.Controls.Add(this.panelHeader);
+            this.Font = new System.Drawing.Font("Segoe UI", 10F);
+            this.Margin = new System.Windows.Forms.Padding(4, 6, 4, 6);
             this.Name = "LoginForm";
-            this.Text = "Přihlášení";
+            this.Text = "Login";
             this.FormClosing += new System.Windows.Forms.FormClosingEventHandler(this.LoginForm_FormClosing);
+            this.panelHeader.ResumeLayout(false);
             this.ResumeLayout(false);
             this.PerformLayout();
+
         }
+
+        private System.Windows.Forms.Label chybaPrihlaseniLabel;
     }
 }